refactor: introduce domain layer with es_de_config and retrodeck_config (#142)
Move es_de_config.py and retrodeck_config.py from lib/ to a new domain/
package, enforcing Cosmic Python layering. Domain has no dependencies on
services, adapters, or lib. Import-linter now enforces 5 contracts.
